Work on AI logic
